requirements.txt
Sphinx==7.3.7
pytest-random-order==1.1.1
allure-pytest==2.13.5